Task number 1 (Main) Notify SF employees in Slack about the release of all new articles on all topics in the company's blog on habr.
Problem number 2 (Additional)
Notify students of individual courses about the release of articles on topics of interest to them. For example, when an article about Python is published: an alert will be sent to the Pythonists 'channel, but not to the Data Scientists' channel.

used ngrok for binding of bot events
used redis+celery
launch manage.py
launch celery and beat (separate terminals)